Squashed commit of the SizeTToLong branch
commit 43ad3cb3a70cc2f784956d9c00f0cec2eab5d133
Author: David <davidkazlauskas@gmail.com>
Date:   Sun Apr 12 21:53:20 2015 +0300

    final touch

commit 8641c66902209a315fb989c15d6760f865640a2f
Author: David <davidkazlauskas@gmail.com>
Date:   Sun Apr 12 21:46:25 2015 +0300

    tests fixed

commit e255786b8fa321fe6b1d61e60ad1bdfe553200e0
Author: David <davidkazlauskas@gmail.com>
Date:   Sun Apr 12 21:41:31 2015 +0300

    checks

commit 05522ab35717e94ded73224ce9c629aa3e78c1dc
Author: David <davidkazlauskas@gmail.com>
Date:   Sun Apr 12 21:34:32 2015 +0300

    biiig changes...

+++ CollectionAdapter.hpp
@@ -57,8 +57,8 @@ struct CollectionAdapter {
     template <class V>
     static void insertAt(ThisCol& c, Iterator at, V&& i);
 
-    static ValueType& getByIndex(ThisCol& c, size_t i);
-    static ConstValueType& getByIndex(ConstCol& c, size_t i);
+    static ValueType& getByIndex(ThisCol& c, long i);
+    static ConstValueType& getByIndex(ConstCol& c, long i);
 
     static long size(ConstCol& c);
 
@@ -67,15 +67,15 @@ struct CollectionAdapter {
 
     static Iterator begin(ThisCol& c);
     static Iterator end(ThisCol& c);
-    static Iterator iterAt(ThisCol& c, size_t i);
+    static Iterator iterAt(ThisCol& c, long i);
 
     static ConstIterator begin(ConstCol& c);
     static ConstIterator end(ConstCol& c);
-    static ConstIterator iterAt(ConstCol& c, size_t i);
+    static ConstIterator iterAt(ConstCol& c, long i);
 
     static ConstIterator cbegin(ConstCol& c);
     static ConstIterator cend(ConstCol& c);
-    static ConstIterator citerAt(ConstCol& c, size_t i);
+    static ConstIterator citerAt(ConstCol& c, long i);
 
     static ValueType& first(ThisCol& c);
     static ConstValueType& first(ConstCol& c);
